refactor: 각 도메인 Model에 대해 DTO 구현
