+++ src/main/kotlin/racingcar/model/car/Cars.kt
@@ -21,13 +21,21 @@ class Cars(_cars: List<Car>) : List<Car> by _cars {
         return this
     }
 
-    fun getWinners(): List<Car> {
+    fun getWinners(): Winners {
         val winnerStandard = this.maxBy { it.position }
-        return this.filter { it.position == winnerStandard.position }
+        return Winners(this.filter { it.position == winnerStandard.position })
     }
+
+    fun toDto(): CarsDto = CarsDto(
+        this.map { car -> car.toDto() }
+    )
 
     companion object {
         private const val DUPLICATED_CAR_NAME_ERROR_MESSAGE =
             "중복된 자동차 이름이 존재합니다."
     }
+}
+
+class CarsDto(_cars: List<CarDto>) : List<CarDto> by _cars {
+    fun toModel(): Cars = Cars(this.map { it.toModel() })
 }
