+++ src/main/kotlin/racingcar/model/round/Round.kt
@@ -1,4 +1,4 @@
-package racingcar.model
+package racingcar.model.round
 
 class Round(val count: Int) {
     init {
@@ -13,4 +13,8 @@ class Round(val count: Int) {
         private const val ROUND_COUNT_OVER_BOUNDARY_ERROR_MESSAGE =
             "라운드 횟수의 범위는 $MIN_ROUND_COUNT 이상 $MAX_ROUND_COUNT 이하입니다."
     }
+}
+
+class RoundDto(private val count: Int) {
+    fun toModel(): Round = Round(count)
 }
