+++ src/test/kotlin/racingcar/model/RoundTest.kt
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Assertions.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.model.round.Round
 import java.lang.IllegalArgumentException
 
 internal class RoundTest {
